requirements.txt
beautifulsoup4==4.8.2
betamax==0.8.1
celery==4.4.0
mypy==0.740
pytest==5.3.1
redis==3.3.11
typing==3.7.4.1
requests
black
Flask==1.1.1
Flask-Babel==0.12.2
flask-blocks==1.0.20
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
python-dotenv==0.10.3
wikipedia==1.4.0
python-igraph
ioahttp